For Python 2, send str as is
For sending a str under Python 2 we used to do an encode() on it, then
send it. The problem was this first does an implicit decode() with the
default encoding (usually ascii) and the decode would fail on anything
that wasn't valid ascii encoded bytes. To solve that we now say that
under Python 2, str types are sent unmodified. For users of the library
it's safest to send strings using the unicode type.