Make it easier to use Connection class directly
Up to now the only supported way to create a new connection was to use
the connect() function. However, some people are using the Connect class
directly and this change makes it a bit easier to do that by making the
class use a contructor which has the same signature as the connect() function.