identifier() should quote if uppercase characters
Previously the identifier() function wouldn't quote the identifier
solely because it contained uppercase characters. The new behaviour of
quoting if there are any uppercase characters is intended to be more in
line with what people expect. It's also easier to recreate the old
behaviour by calling lower() on the input beforehand, than it is to try
and create the new behaviour with the old function.